requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
neo4j>=5.26.0
google-generativeai>=0.8.0
tree-sitter>=0.23.0
tree-sitter-rust>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-go>=0.23.0
pydantic>=2.10.0
httpx>=0.28.0
python-dotenv>=1.0.0
